requirements.txt
Cython
jinja2
pandas
pandas-datareader
pathlib
praw
newspaper3k
requests

